优化修改时的性能，BUG：修改提示重复后，再次双击默认显示是上次修改失败的内容

+++ HuaHaoERP/ViewModel/Customer/CustomerConsole.cs
@@ -9,11 +9,15 @@ namespace HuaHaoERP.ViewModel.Customer
 {
     class CustomerConsole
     {
-        internal bool Add(CustomerModel d)
+        private bool CheckRepeat(CustomerModel d)
         {
             object oTemp;
-            string sql_Repeat = "select 1 from T_UserInfo_Customer where (Number='" + d.Number + "' OR Name='" + d.Name + "') AND DeleteMark IS NULL";
-            if (new Helper.SQLite.DBHelper().QuerySingleResult(sql_Repeat, out oTemp))
+            string sql_Repeat = "select 1 from T_UserInfo_Customer where (Number='" + d.Number + "' OR Name='" + d.Name + "') AND DeleteMark IS NULL AND Guid <> '" + d.Guid + "'";
+            return new Helper.SQLite.DBHelper().QuerySingleResult(sql_Repeat, out oTemp);
+        }
+        internal bool Add(CustomerModel d)
+        {
+            if (CheckRepeat(d))
             {
                 return false;
             }
@@ -26,6 +30,10 @@ namespace HuaHaoERP.ViewModel.Customer
 
         internal bool Update(CustomerModel d)
         {
+            if (CheckRepeat(d))
+            {
+                return false;
+            }
             bool flag = true;
             List<string> sqls = new List<string>();
             string sql_Delete = "Delete From T_UserInfo_Customer Where GUID='" + d.Guid + "'";
